rename LineChartPanel -> GraphQueryProvider, rename LineChart -> LineChartPanel
